typescript(vx-demo): rewrite BarGroup example in TypeScript
