+++ packages/vx-demo/src/components/gallery.js
@@ -15,7 +15,7 @@ import Area from './tiles/area';
 import Stacked from './tiles/stacked';
 import MultiLine from './tiles/multiline';
 import Axis from './tiles/axis';
-import BarGroup from './tiles/bargroup';
+import BarGroup from './tiles/BarGroup.tsx';
 import BarGroupHorizontal from './tiles/bargrouphorizontal';
 import BarStack from './tiles/barstack';
 import BarStackHorizontal from './tiles/barstackhorizontal';
